// src/app/models/rips.ts
export class Rips {
    assignServiceId: number;
    patientId: number;
    documentTypeName: string;
    firstName: string;
    secondName: string;
    surname: string;
    secondSurname: string;
    patientName: string;
    patientDocument: string;
    ripsUserType: string;
    age: number;
    ageUnit: number;
    gender: string;
    providerCode: string;
    businessName: string;
    adomIdentificationType: string;
    adomIdentificationNumber: string;
    departmentCode: string;
    cityCode: string;
    residenceArea: string;
    authorizationNumber: string;
    contractNumber: string;
    serviceId: number;
    serviceName: string;
    entityId: number;
    rate: number;
    cups: string;
    cie10: number;
    external: number;
    consultation:number;
    entityCode: string;
    entityName: string;
    quantity: number;
    quantityCompleted: number;
    initialDate: string;
    finalDate: string;
    professionalId: number;
    professionalName: string;
    coPaymentAmount: number;
    coPaymentFrecuencyId: number;
    coPaymentFrecuencyName: string;
    stateId: number;
    stateName: string;
    totalRows: number;
    copaymentStatus: string;
    totalCopaymentReported: number;
    otherValuesReported: number;
    totalCopaymentReceived: number;
    discounts: number;
    valueToPayToProfessional: number;
    grandTotalToPay: number;
    planEntityName: string;
    invoiceNumber: string;
}